### 2018-09-20 cbaca89
added space to ENJ affiliations in paper.md
