### 2018-09-20 a9e58ac
added affiliation index in paper.md
