### 2018-09-20 #963 @jeffmacinnes
Hi @cMadan! I fixed the PDF issue, and you can find the proof above. From looking through the list of reviewers, I think @shivam11, @nnadeau, and @ionlights would all be well-qualified reviewers with the expertise to assess our methods and the contribution of our package. Please let me know if there's anything else I can do to help. Cheers!
