### 2018-09-20 #963 @cMadan
@jeffmacinnes, thanks for sorting that out! It looks like your license is a BSD-3 clause license, but has information about a related patent at the top. I'm not sure if that still fits JOSS' requirements re: using an OSI-approved license, so I'm going to consult with the other JOSS editors and get back to you.

@nnadeau, assuming the license issue get sorted, it would be great to have you as a reviewer!

@shivam11 @ionlights, please let me know if you're also able to review this submission.
