### 2018-09-20 ad2bf80
Fixed file formats in readme folder structure
