### 2018-09-20 #963 @jeffmacinnes
@cMadan Preference would be for the MIT license, but our package makes use of the SIFT algorithm (as part of the [OpenCV](https://docs.opencv.org/3.4.3/da/df5/tutorial_py_sift_intro.html) library), which is patented for any commercial use. I included the current license file to address that, but welcome any other suggestions!
