### 2018-09-20 #963 @cMadan
@jeffmacinnes, I discussed this wither other JOSS editors and we decided that the license needs to be an OSI-approved license without modification (https://opensource.org/licenses/alphabetical), and that an OSI+Patents license is not acceptable.

I understand that SIFT is a standard CV method, so I'm not quite sure what to do here. I'll ask and see if any of the other editors can comment here and give you more specific advice.
