### 2018-09-23 #963 @jeffmacinnes
@cMadan, gotcha, that's a reasonable policy. I noticed that OpenCV, as well as the [opencv_contrib](https://github.com/opencv/opencv_contrib) extra modules, where SIFT methods are found, are distributed with unmodified 3-clause BSD licenses:

[opencv license](https://github.com/opencv/opencv/blob/master/LICENSE)
[opencv_contrib license](https://github.com/opencv/opencv_contrib/blob/master/LICENSE)

Would a similar strategy work for our package? Thanks for your help in sorting this out.
