### 2018-09-26 #963 @cMadan
Yes, I think that should be fine here. How about you update your license file and I'll then run by the other editors to confirm, though I think that's the best you can do.
