### 2018-09-27 989ae80
updated license to BSD-3 clause
