+++ README.md
@@ -1,4 +1,4 @@
-[![Build Status](https://travis-ci.com/jeffmacinnes/mobileGazeMapping.svg?branch=master)](https://travis-ci.com/jeffmacinnes/mobileGazeMapping)
+[![License](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause) [![Build Status](https://travis-ci.com/jeffmacinnes/mobileGazeMapping.svg?branch=master)](https://travis-ci.com/jeffmacinnes/mobileGazeMapping)
 
 # Mobile Gaze-Mapping
 
