+++ LICENSE.txt
@@ -1,57 +1,11 @@
 Copyright (c) 2018, Jeff MacInnes
-All rights reserved.
 
-The following patent has been issued for methods embodied in this
-software: "Method and apparatus for identifying scale invariant features
-in an image and use of same for locating an object in an image," David
-G. Lowe, US Patent 6,711,293 (March 23, 2004). Provisional application
-filed March 8, 1999. Asignee: The University of British Columbia. For
-further details, contact David Lowe (lowe@cs.ubc.ca) or the
-University-Industry Liaison Office of the University of British
-Columbia.
+Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
 
-Note that restrictions imposed by this patent (and possibly others)
-exist independently of and may be in conflict with the freedoms granted
-in this license, which refers to copyright of the program, not patents
-for any methods that it implements.  Both copyright and patent law must
-be obeyed to legally use and redistribute this program and it is not the
-purpose of this license to induce you to infringe any patents or other
-property right claims or to contest validity of any such claims.  If you
-redistribute or use the program, then this license merely protects you
-from committing copyright infringement.  It does not protect you from
-committing patent infringement.  So, before you do anything with this
-program, make sure that you have permission to do so not merely in terms
-of copyright, but also in terms of patent law.
+1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
 
-Please note that this license is not to be understood as a guarantee
-either.  If you use the program according to this license, but in
-conflict with patent law, it does not mean that the licensor will refund
-you for any losses that you incur if you are sued for your patent
-infringement.
+2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are
-met:
-    * Redistributions of source code must retain the above copyright and
-      patent notices, this list of conditions and the following
-      disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-      notice, this list of conditions and the following disclaimer in
-      the documentation and/or other materials provided with the
-      distribution.
-    * Neither the name of Duke University, the University of Washington,
-      the University of Pennsylvania, nor the names of its contributors may be
-      used to endorse or promote products derived from this software without
-      specific prior written permission.
+3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
 
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
[…]
