### 2018-09-27 #963 @jeffmacinnes
Great, thanks @cMadan -- just updated the repo with the new license
