### 2018-09-28 #984 @whedon
[REVIEW]: Mobile Gaze Mapping: A Python package for mapping mobile gaze data to a fixed target stimulus
**Submitting author:** @jeffmacinnes (<a href="http://orcid.org/0000-0003-2388-0699">Jeffrey MacInnes</a>)
**Repository:** <a href="https://github.com/jeffmacinnes/mobileGazeMapping" target ="_blank">https://github.com/jeffmacinnes/mobileGazeMapping</a>
**Version:** v1.0.0
