### 2018-10-17 2774c95
fixed typos
