### 2018-10-19 #984 @cMadan
@nnadeau, during the pre-review (https://github.com/openjournals/joss-reviews/issues/963) I noticed some typos in the README, so I fixed them (https://github.com/jeffmacinnes/mobileGazeMapping/pull/1).
