### 2018-10-26 #984 @zablab
> @jeffmacinnes how are your co-authors involved in the repo? They have no commits.

I am one of the co-authors. Trying to streamline correspondence, so letting @jeffmacinnes take the lead--just FYI he's been out of the country for the past week.
