### 2018-10-27 1030e0e
fixed submitted issues
