+++ preprocessing/pl_preprocessing.py
@@ -5,16 +5,18 @@ Tested with Python 3.6, open CV 3.2
 The output directory will be created within the output root path specified by the user, and be named according to [mo-day-yr]/[hr-min-sec] of the original creation time.
 
 The output directory will contain:
-	- worldCamera.mp4: the video from the point-of-view scene camera on the glasses
-	- frame_timestamps.tsv: table of timestamps for each frame in the world
-	- gazeData_world.tsv: gaze data, where all gaze coordinates are represented w/r/t the world camera
+    - worldCamera.mp4: the video from the point-of-view scene camera on the glasses
+    - frame_timestamps.tsv: table of timestamps for each frame in the world
+    - gazeData_world.tsv: gaze data, where all gaze coordinates are represented w/r/t the world camera
 """
 
 # python 2/3 compatibility
 from __future__ import division
 from __future__ import print_function
 
-import sys, os, shutil
+import os
+import sys
+import shutil
 import argparse
 from datetime import datetime
 from os.path import join
@@ -27,156 +29,151 @@ import gc
 import msgpack
 
 def preprocessData(inputDir, output_root):
-	"""
-	Run all preprocessing steps for pupil lab data
-	"""
-	### Prep output directory
-	info_file = join(inputDir, 'info.csv')  	# get the timestamp from the info.csv file
-	with open(info_file, 'r') as f:
-		for line in f:
-			if 'Start Date' in line:
-				startDate = datetime.strptime(line.split(',')[1].strip('\n'), '%d.%m.%Y')
-				date_dir = startDate.strftime('%Y_%m_%d')
-			if 'Start Time' in line:
-				startTime = datetime.strptime(line.split(',')[1].strip('\n'), '%H:%M:%S')
-				time_dir = startTime.strftime('%H-%M-%S')
-			if 'World Camera Resolution' in line:
-				worldCamRes_y = int(line.split(',')[1].split('x')[1].strip('\n'))
+    """ Run all preprocessing steps for pupil lab data """
+    ### Prep output directory
+    info_file = join(inputDir, 'info.csv')      # get the timestamp from the info.csv file
+    with open(info_file, 'r') as f:
+        for line in f:
+            if 'Start Date' in line:
+                startDate = datetime.strptime(line.split(',')[1].strip('\n'), '%d.%m.%Y')
+                date_dir = startDate.strftime('%Y_%m_%d')
+            if 'Start Time' in line:
+                startTime = datetime.strptime(line.split(',')[1].strip('\n'), '%H:%M:%S')
+                time_dir = startTime.strftime('%H-%M-%S')
 
-	# create the output directory (if necessary)
-	outputDir = join(output_root, date_dir, time_dir)
-	if not os.path.isdir(outputDir):
-		os.makedirs(outputDir)
+    # create the output directory (if necessary)
+    outputDir = join(output_root, date_dir, time_dir)
+    if not os.path.isdir(outputDir):
+        os.makedirs(outputDir)
 
-	### Format the gaze data
-	print('formatting gaze data...')
-	gazeData_world, frame_timestamps = formatGazeData(inputDir)
+    ### Format the gaze data
+    print('formatting gaze data...')
+    gazeData_world, frame_timestamps = formatGazeData(inputDir)
 
-	# write the gazeData to to a csv file
-	print('writing file to csv...')
-	csv_file = join(outputDir, 'gazeData_world.tsv')
-	export_range = slice(0, len(gazeData_world))
-	with open(csv_file, 'w', encoding='utf-8', newline='') as csvfile:
-		csv_writer = csv.writer(csvfile, quoting=csv.QUOTE_NONE)
[…]
