+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+
+Important changes and updates to the code will be documented in this file. This packages uses [https://calver.org/](https://calver.org/) versioning
+
+## [18.10.01]
+### Added
+- This CHANGELOG file to keep track of modifications to the repository
+- `/docs` dir to hold all documentation files
+- `CONTRIBUTING.md` guidelines added to `docs` dir
+- `ISSUE_TEMPLATE.md` added to `docs` dir
+### Fixed
+- cleaned all files to remove unused code fragments
+- added `pytest` to `requirements.txt` file 
