+++ requirements.txt
@@ -1,3 +1,4 @@
 numpy>=1.13.2
 pandas>=0.19.2
 opencv-contrib-python==3.4.2.17
+pytest>=3.8.0
