### 2018-10-27 1154ac9
updated automated tests to provide more thorough testing on outputs
