+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 Important changes and updates to the code will be documented in this file. This packages uses [https://calver.org/](https://calver.org/) versioning
 
-## [18.10.01]
+## [2018.10.02]
+### Changed
+- `tests/test_mapGaze.py` modified to run more thorough tests and confirmation on output values
+
+## [2018.10.01]
 ### Added
 - This CHANGELOG file to keep track of modifications to the repository
 - `/docs` dir to hold all documentation files
@@ -10,4 +14,4 @@ Important changes and updates to the code will be documented in this file. This 
 - `ISSUE_TEMPLATE.md` added to `docs` dir
 ### Fixed
 - cleaned all files to remove unused code fragments
-- added `pytest` to `requirements.txt` file 
+- added `pytest` to `requirements.txt` file
