### 2018-10-27 e3743ef
changed test file assertions to almost_equal to account for platform differences in floats
