### 2018-10-27 e0fad18
changed test file assertions to allclose to account for rounding differences
