### 2018-10-27 #984 @jeffmacinnes
@nnadeau This package is the culmination of a project that my co-authors and I developed and tested over the course of a year. However, it was only recently that we made the decision to modify the tools so that we may release publicly via github and JOSS. The repository only reflects those recent changes.
