### 2018-11-12 1e7b0f1
updated CHANGELOG to reflect merged pull requests re typos on preprocessing scripts
