### 2018-11-17 #984 @jeffmacinnes
Hi @nnadeau, I noticed two pull requests that I had overlooked earlier this week, so went ahead and merged -- thanks for those fixes! I think that we've addressed the other concerns, but please let me know if there's anything else that I'm missing.
