### 2018-11-17 jeffmacinnes/mobileGazeMapping#13 @nnadeau
Output file formats do not match README
* Date you used mobileGazeMapping: 2018/11/17
* mobileGazeMapping version used: master (`1e7b0f15cefcda3f8011aa70bd20c674fec71e83`)
* Python version: Python 3.6.6
* Operating System: Ubuntu 18.04.1 LTS

### Description

- Output files do not match README
- I get `m4v` files, not `mp4`
- Interestingly, the output matches what was original before https://github.com/jeffmacinnes/mobileGazeMapping/commit/ad2bf800b098a8ec32c12dbdfae053cde62c108d#diff-04c6e90faac2675aa89e2176d2eec7d8

### What You Did

- Ran:

```
python mapGaze.py tests/gazeData_world.tsv tests/worldCamera.mp4 tests/referenceImage.jpg
```

- Output:
```
tests/mappedGazeOuput/
├── gazeData_mapped.tsv
├── mapGazeLog.log
├── ref2world_mapping.m4v
├── ref_gaze.m4v
├── referenceImage.jpg
└── world_gaze.m4v
```
