### 2018-11-17 jeffmacinnes/mobileGazeMapping#14 @nnadeau
Missing DOIs
- [docs/paper/paper.bib](https://github.com/jeffmacinnes/mobileGazeMapping/blob/master/docs/paper/paper.bib) is missing DOIs for certain references
- e.g., http://www.scitepress.org/DigitalLibrary/Link.aspx?doi=10.5220/0001787803310340
- Use cross-ref: https://www.crossref.org/
