### 2018-11-17 #984 @nnadeau
@jeffmacinnes last two issues:
- https://github.com/jeffmacinnes/mobileGazeMapping/issues/14
- https://github.com/jeffmacinnes/mobileGazeMapping/issues/13

Then I believe it's all good to go! Kudos!
