### 2018-11-19 0e958ed
added dois to bibtex refs, fixed video file extensions in documentation
