+++ docs/paper/paper.bib
@@ -1,19 +1,28 @@
-@article{Lowe04distinctiveimage,
-    author = {David G. Lowe},
-    title = {Distinctive Image Features from Scale-Invariant Keypoints},
-    journal = {International Journal of Computer Vision},
-    year = {2004},
-    volume = {60},
-    pages = {91--110}
+@Article{Lowe2004,
+    author = "Lowe, David G.",
+    title = "Distinctive Image Features from Scale-Invariant Keypoints",
+    journal = "International Journal of Computer Vision",
+    year = "2004",
+    month = "Nov",
+    day = "01",
+    volume = "60",
+    number = "2",
+    pages = "91--110",
+    abstract="This paper presents a method for extracting distinctive invariant features from images that can be used to perform reliable matching between different views of an object or scene. The features are invariant to image scale and rotation, and are shown to provide robust matching across a substantial range of affine distortion, change in 3D viewpoint, addition of noise, and change in illumination. The features are highly distinctive, in the sense that a single feature can be correctly matched with high probability against a large database of features from many images. This paper also describes an approach to using these features for object recognition. The recognition proceeds by matching individual features to a database of features from known objects using a fast nearest-neighbor algorithm, followed by a Hough transform to identify clusters belonging to a single object, and finally performing verification through least-squares solution for consistent pose parameters. This approach to recognition can robustly identify objects among clutter and occlusion while achieving near real-time performance.",
+    issn = "1573-1405",
+    doi = "10.1023/B:VISI.0000029664.99615.94",
+    url = "https://doi.org/10.1023/B:VISI.0000029664.99615.94"
 }
-@inproceedings{muja_flann_2009,
-  author    = {Marius Muja and David G. Lowe},
-  title     = {Fast Approximate Nearest Neighbors with Automatic Algorithm Configuration},
-  booktitle = {International Conference on Computer Vision Theory and Application
-               VISSAPP'09)},
-  publisher = {INSTICC Press},
-  year      = {2009},
-  pages     = {331-340}
+@conference{visapp09,
+    author = {Marius Muja and David G. Lowe},
+    title = {FAST APPROXIMATE NEAREST NEIGHBORS WITH AUTOMATIC ALGORITHM CONFIGURATION},
+    booktitle = {Proceedings of the Fourth International Conference on Computer Vision Theory and Applications - Volume 1: VISAPP, (VISIGRAPP 2009)},
+    year = {2009},
+    pages = {331-340},
+    publisher = {SciTePress},
+    organization = {INSTICC},
+    doi = {10.5220/0001787803310340},
+    isbn = {978-989-8111-69-2},
 }
 @article{opencv_library,
     author = {Bradski, G.},
