### 2018-11-19 jeffmacinnes/mobileGazeMapping#13 @jeffmacinnes
Thanks for catching that. I fixed the README file (as well as the comments and documentation in mapGaze.py to reflect the correct file extensions
