### 2018-11-19 jeffmacinnes/mobileGazeMapping#14 @jeffmacinnes
I added DOIs to all references except the one for OpenCV. For that one, I follow their recommended bibTex citation at: [https://github.com/opencv/opencv/wiki/CiteOpenCV](https://github.com/opencv/opencv/wiki/CiteOpenCV)
