### 2018-11-19 #984 @jeffmacinnes
Thanks @nnadeau! I just fixed those last two issues. I really appreciate your help reviewing this repo and helping us get the package prepped for release. @cMadan -- just let us know what's next!
