### 2018-11-20 #984 @cMadan
@nnadeau, can you check and see if you're satisfied with those issues? If so, please update the reviewer checklist at the top.

@jeffmacinnes, after @nnadeau confirms everything is all set, I'll look over things one last time and see if there are any minor issues remaining. If not, I'll ask you to archive the code in Zenodo or figshare and you'll be all set! So, just sit tight for now, but we're almost done.
