### 2018-11-22 #984 @cMadan
Thank you for the thorough review, @nnadeau!

@jeffmacinnes, everything looks good to me! Now all I need is the doi for an archived version of the code (i.e., on Zenodo or figshare).
