### 2018-11-22 3c0c925
added DOI badge to README
