### 2018-11-22 4a9997e
added DOI badge to README
