+++ README.md
@@ -1,4 +1,5 @@
-[![License](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause) [![Build Status](https://travis-ci.com/jeffmacinnes/mobileGazeMapping.svg?branch=master)](https://travis-ci.com/jeffmacinnes/mobileGazeMapping) ![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1494278.svg)
+[![License](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause) [![Build Status](https://travis-ci.com/jeffmacinnes/mobileGazeMapping.svg?branch=master)](https://travis-ci.com/jeffmacinnes/mobileGazeMapping) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1494278.svg)](https://doi.org/10.5281/zenodo.1494278)
+
 
 # Mobile Gaze-Mapping
 
